moved all tests from nitrogen
